Handle multiple feeds and return only unique posts
When adding to the `feed` parameter multiple feed urls the results
would be from the default configuration value. The reason is because
the feed urls were being read as a single string

```
"feed_url_1,feed_url_2"
```
A fix was breaking them into an array.

```
["feed_url_1", "feed_url_2"]
```
This change will also fix the templates from repeated posts.